wallet-responive
Separated the wallet connection

Will call for it in header if above 767px width

Will call for it in footer if below 768px width. Sticky in bottom right for mobile
