+++ src/ui/components/common.js
@@ -11,6 +11,40 @@ import { paneStyles, colStyles } from '../components/styles'
 import { getTokenSymbol } from '../../client/web3'
 import { H1, HR, Colour, Text, Center, Label, Sublabel, LabelGroup } from '../components/elements'
 
+// Check If Responsive
+export const useWindowSize = () => {
+  // Initialize state with undefined width/height so server and client renders match
+  const [windowSize, setWindowSize] = useState({
+    width: undefined,
+    height: undefined,
+  });
+
+  useEffect(() => {
+    // Handler to call on window resize
+    function handleResize() {
+      // Set window width/height to state
+      setWindowSize({
+        width: window.innerWidth,
+        height: window.innerHeight,
+      });
+    }
+
+    // Add event listener
+    window.addEventListener("resize", handleResize);
+
+    // Call handler right away so state gets updated with initial window size
+    handleResize();
+
+    // Remove event listener on cleanup
+    return () => window.removeEventListener("resize", handleResize);
+  }, []); // Empty array ensures that effect is only run on mount
+
+  return windowSize;
+}
+
+
+
+
 export const BreadcrumbCombo = (props) => {
 
   return (
