+++ src/ui/layout/Headbar.js
@@ -1,110 +1,21 @@
-import React, { useState, useEffect, useContext } from 'react'
-import { Context } from '../../context'
+import React from 'react'
 import { Link } from "react-router-dom";
-import { Row, Col, Layout, Menu, Drawer } from 'antd';
-import { UserOutlined, LoadingOutlined } from '@ant-design/icons';
+import { Row, Col, Layout, Menu } from 'antd';
 import '../../App.css';
-import { Button } from '../components/elements'
 
-import Web3 from 'web3'
 // import axios from 'axios'
 
-import { message } from 'antd';
 
 // import logo from '../../assets/spartan-logo-white.png';
-
-import WalletDrawer from './WalletDrawer'
-import { getAddressShort, } from '../../utils'
-import {
-    getTokenDetails, getListedTokens,
-    getWalletData, getStakesData, getListedPools
-} from '../../client/web3'
+import AddressConn from '../components/AddressConn'
+import { useWindowSize } from '../components/common'
 
 const { Header } = Layout;
 
 const Headbar = (props) => {
 
-    const context = useContext(Context)
-    const [connecting, setConnecting] = useState(false)
-    const [connected, setConnected] = useState(false)
-    const [visible, setVisible] = useState(false);
-
-    useEffect(() => {
-        connectWallet()
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [])
-
-    const connectWallet = async () => {
-        setConnecting(true)
-        window.web3 = new Web3(window.ethereum);
-        const account = (await window.web3.eth.getAccounts())[0];
-        if (account) {
-            message.loading('Loading tokens', 3);
-            // let assetArray = context.assetArray ? context.assetArray : await getAssets()
-            // context.setContext({ 'assetArray': assetArray })
-            // let assetDetailsArray = context.assetDetailsArray ? context.assetDetailsArray : await getTokenDetails(account, assetArray)
-            // context.setContext({ 'assetDetailsArray': assetDetailsArray })
-
-            let tokenArray = context.tokenArray ? context.tokenArray : await getListedTokens()
-            context.setContext({ 'tokenArray': tokenArray })
-            // context.setContext({ 'poolsData': await getPoolsData(tokenArray) })
-
-            // let allTokens = assetArray.concat(tokenArray)
-            // var sortedTokens = [...new Set(allTokens)].sort()
-
-            let tokenDetailsArray = context.tokenDetailsArray ? context.tokenDetailsArray : await getTokenDetails(account, tokenArray)
-            context.setContext({ 'tokenDetailsArray': tokenDetailsArray })
-
-            message.loading('Loading wallet data', 3);
-            let walletData = await getWalletData(account, tokenDetailsArray)
-            context.setContext({ 'walletData': walletData })
-
-            let poolArray = context.poolArray ? context.poolArray : await getListedPools()
-            context.setContext({ 'poolArray': poolArray })
-
-            let stakesData = context.stakesData ? context.stakesData : await getStakesData(account, tokenArray)
-            context.setContext({ 'stakesData': stakesData })
-
-            context.setContext({ 'connected': true })
-            await getSpartaPrice()
-            setConnecting(false)
-            setConnected(true)
-            message.success('Loaded!', 2);
-        } else {
-            await ethEnabled()
-            setConnected(false)
-        }
-    }
-
-    const ethEnabled = () => {
-        console.log('connecting')
-        if (window.ethereum) {
-            window.web3 = new Web3(window.ethereum);
-            window.ethereum.enable();
-            setConnecting(true)
-            return true;
-        }
-        return false;
-    }
-
-    const getSpartaPrice = async () => {
-        // let resp = await axios.get('https://api.coingecko.com/api/v3/simple/price?ids=spartan&vs_currencies=usd')
-        // console.log(resp.data.spartan.usd)
-        // context.setContext({ 'spartanPrice': resp.data.spartan.usd })
-        context.setContext({ 'spartanPrice': 0.3 })
-        return
-    }
-
-    const addr = () => {
-        return getAddressShort(context.walletData?.address)
-    }
-
-    const showDrawer = () => {
-        setVisible(true);
-    };
-    const onClose = () => {
-        setVisible(false);
-    };
+  const size = useWindowSize();
+  const width = size.width;
 
     return (
         <Header>
@@ -150,28 +61,12 @@ const Headbar = (props) => {
                         </Menu.Item> */}
                     </Menu>
                 </Col>
-                <Col xs={4} style={{ textAlign: 'right' }}>
-                    {!connected && !connecting &&
-                        <Button type="primary" onClick={connectWallet}>CONNECT</Button>
-                    }
-                    {connecting &&
-                        <Button type="primary" icon={<LoadingOutlined />}>CONNECTING</Button>
-                    }
-                    {connected &&
-                        <Button type="primary" icon={<UserOutlined />} onClick={showDrawer}>{addr()}</Button>
-                    }
-                </Col>
+                { width > 767 &&
+                  <Col xs={0} md={4} style={{ textAlign: 'right' }}>
+                      <AddressConn />
+                  </Col>
+                }
             </Row>
-            <Drawer
-                title={context.walletData?.address}
-                placement="right"
-                closable={false}
-                onClose={onClose}
-                visible={visible}
-                width={350}
-            >
-                <WalletDrawer />
-            </Drawer>
 
         </Header>
     )
